requirements.txt
git+https://github.com/msuliot/package.utils.git
nltk
pymongo
openai
pinecone-client